Trim legacy.cpp and legacy.h
