+++ src/clkConformationBfacf3.cpp
@@ -9,6 +9,7 @@
 #include "legacyBfacf.h"
 #include "threevector.h"   //Diwen 08/22/2018
 #include "legacy.h" //Diwen 01/08/2019
+#include "writhe.h"
 
 #include <cmath> // Diwen 01/07/2019
 #include <cstdlib>
