+++ src/clkConformationSeqconvert.cpp
@@ -9,6 +9,7 @@
 
 #include "clkConformationSeqconvert.h"
 #include "legacy.h"
+#include "writhe.h"
 
 using namespace std;
